Fix [EI-758] Fault sequence is not executed when on complete is executed
aggregate messages based on timeout and gave an invalid sequence to onComplete
tag. For that scenario, fault sequence is NOT executed. Reason is timeout
process is carried out by a different thread and it does not execute the
fault sequence. Fix is to get fault stack and invoke relevant fault sequence.

Issue: https://github.com/wso2/product-ei/issues/758