Fix wso2/product-ei/issues/1719
In EndPoint error handling scenarios when we define the
“responseAction” as “discard” the endpoint will not be get suspended
whether the “retriesBeforeSuspension” count exceeded or not.
This has been fixed by changing the logic to execute the onFault
method of AbstractEndpoint wheteher the timeOutAction defined
as Discard or Fault and deciding whether to invoke the fault
sequence or not in the onFault method of AbstractEndpoint.

Fixes: https://github.com/wso2/product-ei/issues/1719