Fix wso2/product-ei/issues/1054
When deploying synapse artifacts with the given configurations,
AbstractMediatorFactory tries to create a new ScriptMediator. When
creating the ScriptMediator it tries to initialize the ScriptEngine.
In this method it calls the getEngineByExtension method of bsf library
but since the groovy library is missing it gives exception.
But getEngineByExtension method haven't handled the exception.
To fix this a SynapseArtifactDeploymentException have been
thrown in the finally block.

Fixes: https://github.com/wso2/product-ei/issues/1054