Fix wso2/product-ei/issues/1054
When deploying synapse artifacts with the given configurations,
AbstractMediatorFactory tries to create a new ScriptMediator. When
creating the ScriptMediator it tries to initialize the ScriptEngine.
In this method it calls the getEngineByExtension method of bsf library
but since the groovy library is missing it throws NoClassDefFoundError.
But deploy method of AbstractSynapseArtifactDeployer haven't handled
this error. So service get deployed without ScriptMediator.
To fix this a NoClassDefFoundError have been
caught in the deploy method of AbstractSynapseArtifactDeployer.

Fixes: https://github.com/wso2/product-ei/issues/1054