Fix wso2/product-ei/issues/1820
When a proxy with groove dependency is added to the EI-6.1.1 then the
server is started but the mgt console is not accessible.
This is happening since when calliing the getEngineByExtension method
of bsf library it throws NoClassDefFoundError and it haven't been
handled. To fix this NoClassDefFoundError have been
caught in the SynapseXMLConfigurationFactory.

Fixes: https://github.com/wso2/product-ei/issues/1820