Fix Dropping headers over local transport
Content-Type header of GET calls is getting dropped if the preceded call is a data service call over local transport. This was due to using hashmaps instead of Treemaps for transport headers in property and header mediators. When creating the target request we check whether the transport header object type as a treemap.

This will fix the issue by using Treemaps and also checking Map type instead of limiting to specific map implementation.

Resolves wso2/product-ei#2143