Updated removeSeqContinuationState method in ContinuationStackManager
For anonymous sequences, the reliant continuation state is removed (as
opposed to popping the continuation state stack) after processing. This is an error,
since no reliant continuation state is added when the anonymous sequence
is started: the method addSeqContinuationState only pushes a
continuation state onto the stack for non-anonymous sequences, but
doesn't do anything for anonymous sequences.