Fix messageType not being set correctly
Fixes wso2/product-apim#3537

While sending response back to client in JMS dual channel scenario,
we use the messageType set in the callback, not the one that is set
in the response msg context. This causes responding to client with
incorrect message type (form-urlencoded). This fix will use the
content type sent from the backend as the message type and set this
value to response msg context.