Fix wso2/product-ei/issues/4961
When the socket timeout exceeds in the REQUEST_DONE state
TargetHandler invokes the targetErrorHandler and it invokes
the SynapseCallbackRecievers receive method. But when this
happens the call back is already removed from the map. So
the SynapseCallbackReciver prints the "Synapse received a
response for the request with ...
But a callback is not registered (anymore) to process this response"
log.
This can occur at any time where targetErrorHandler invokes the
 SynapseCallbackRecievers receive method.
Fix this issue by differentiating the path by adding a state when
SynapseCallbackReciever is invoked by targetErrorHandler

Fixes: https://github.com/wso2/product-ei/issues/4961