Fix wso2/product-ei/issues/4992
After refactoring the ForwardingService the dispatching message
has moved to a new method. Previously when trying to dispatch the message
as the first step of the while loop it has cloned and set the original envelope
to the dispatching message. But after previously mentioned refactoring rather than
using the original envelope it has used the envelope passed as method argument.
But since this reference is getting allocated with response in the dispathing method
It has retried with the response message context.
This is fixed by setting the original envelope rather than using argument envelope.

Fixes: https://github.com/wso2/product-ei/issues/4992