Fix https://github.com/wso2/micro-integrator/issues/1739
When original store is down, the messages are moved to failover
ms in failover msmp scenario. But when again the original stroe is
connected (broker) failover processor is unable to send to original
store since it uses the existing producer of original store since it
hasn't been assigned to null. So solve this by assigning null to
store producer when connection is not available.

Fixes: https://github.com/wso2/micro-integrator/issues/1739