Allow empty payload without FORCE_POST_PUT_NOBODY property
The FORCE_POST_PUT_NOBODY property is required when sending an empty payload to
the backend with the POST or PUT request. The usage of this property has been removed
and modify the transport. There was an issue with the GET request in service chaining.
The entire message body is not reading before sending the request to the backend.
It is fixed with this PR as well.
Resolves: https://github.com/wso2/product-ei/issues/4376