Fix the importing custom mediators issue in the integration studio
Custom Mediators are considered as comment mediators to be able to
import the sequences with custom mediators. When deserializing
custom mediators which starts as "CUSTOM_" is considered as
comment mediators and is added to the mediators sequence.
This fix will treat the comment mediators starting with the comment
text of "<CUSTOM_" as custom mediators and will add the comment
text ( custom mediator) as a child to the OM parent.