Move request message discard logic from TargetHandler to ClientWorker
Previously, in error scenarios, the Passthrough Pipe was consumed and discarded in the TargetHandler.
As a result, the HTTPS-Sender I/O dispatcher thread will get blocked until the timeout.
Therefore the discard logic is moved to the ClientWorker, in which a PassThroughMessageProcessor thread will be used.

Git issue: https://github.com/wso2/api-manager/issues/1470