Close target connection upon endpoint timeout
In the current architecture, when the endpoint timeouts the timeout handler will call the fault sequence and finish the mediation flow for that request.
But the connection to the backend is not closed until the socket timeout.

This fix will improve the architecture to close the socket connection when an endpoint timeout happens.

Git issue: https://github.com/wso2/api-manager/issues/1746