Fix input read failure when connection is reused
Issue here is when there is an exact amount of bytes equal to the buffer
(or Greater than the buffer), in the producePostActions() method we call suspendInput()
in the underlying connection.
So when we write back bytes to the source, we are supposed to reset this
flag in consumePostActions() method to free the connection to use for consuming backend input.
But in the first case (bytes are exactly equal to buffer size) we are not free up the connection.
So as the solution lift the event mask when releasing the connection

Fixes: https://github.com/wso2/api-manager/issues/1779