Use endpoint URL and OAuth credentials when caching token
Since the cached token was mapped to the synapse endpoint instance previously, the same
token was shared even we change the endpoint URL dynamically. With this improvement, we
use the endpoint URL and OAuth credentials as the key when caching the token.

Git issue: https://github.com/wso2/api-manager/issues/2381