Disable default access logger and utilize log4j2 for http access logs
Related to https://github.com/wso2/micro-integrator/issues/3281
Currently, HTTP Access logs use a custom implementation to write the logs to the file. This has a request and response queue to hold the logs and write them to the file using a Timer task. In a high-traffic scenario, this becomes a bottleneck because the Heap memory easily fills up because of the logs added to the queuea.

This commit disables the current implementation and utilizes log4j2 to write the access logs to the file system which is more efficient. Also, timer tasks are removed and LinkedBlockingQueue is used to handle logs asynchronously in a timely manner.