Fix CPU spinning issue
When back-end connection is closed (RST sent) while handling the
response if the protocol state is already marked as RESPONSE_DONE in the
TargetHandler exception patch we have omitted to close the connection so the connection
remain intact and it will continue to hit the inputReady method and
when trying to read from the SSLIOSession it will throws a I/O Exception
but since in the exception patch we are omitting the error this will
loop.
Fixed this issue by moving the connection close logic to finally block.

Fixes: https://github.com/wso2/micro-integrator/issues/3363